Implemented scaling factor and score system for creatures

+++ simulation.cpp
@@ -23,12 +23,12 @@ void Simulation::setGravity(float x, float y) {
     world.SetGravity(b2Vec2(x, y));
 }
 
-void Simulation::createMob(string filePath, int posX, int posY, int sizeX, int sizeY) {
+void Simulation::createMob(string filePath, float posX, float posY, float sizeX, float sizeY) {
     Mob* newMob = new Mob(filePath, posX, posY, sizeX, sizeY, world);
     genericMobs.push_back(newMob);
 }
 
-void Simulation::createMob(string filePath, int posX, int posY, int sizeX, int sizeY, string name) {
+void Simulation::createMob(string filePath, float posX, float posY, float sizeX, float sizeY, string name) {
     Mob* newMob = new Mob(filePath, posX, posY, sizeX, sizeY, world);
     namedMobs.insert(pair<string, Mob*>(name, newMob));
 }
